Reorganize experiments
