+++ benchmark/experiments/toy/generate_integrator_cycle_plots.py
